Day 2v2

+++ Tresor001/Product.cs
@@ -17,7 +17,7 @@ namespace Tresor001
         }
 
         public string product_description { get; set; }
-        public int product_rating { get; set; }
+        public double product_rating { get; set; }
         public string product_name { get; set; }
     }
 }
