+++ Tresor001/Review.cs
@@ -15,7 +15,7 @@ namespace Tresor001
             PartitionKey = partition;
             RowKey = row;
         }
-        public int review_rating { get; set; }
+        public double review_rating { get; set; }
         public string review_name { get; set; }
         public string review_text { get; set; }
         public string review_category { get; set; }
